requirements.txt
flask
streamlit
nltk
textblob
scikit-learn
numpy
pandas
requests
python-dotenv
plotly
spacy